+++ src/main/kotlin/racingcar/Application.kt
@@ -1,5 +1,7 @@
 package racingcar
 
+import racingcar.controller.RaceController
+
 fun main() {
     RaceController().run()
 }
