--- src/main/kotlin/racingcar/GenerateRandomNumber.kt
@@ -1,5 +1,0 @@
-package racingcar
-
-class GenerateRandomNumber {
-    fun randomNumber() = (0..9).random()
-}
