+++ src/main/kotlin/racingcar/domain/GenerateRandomNumber.kt
@@ -1,0 +1,7 @@
+package racingcar.domain
+
+import racingcar.utils.ErrorHandler.validRandomNumber
+
+class GenerateRandomNumber {
+    fun randomNumber() = (0..9).random().validRandomNumber()
+}
